% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.vdoe.vpi-package.R
\docType{package}
\name{dc.vdoe.vpi-package}
\alias{dc.vdoe.vpi}
\alias{dc.vdoe.vpi-package}
\title{dc.vdoe.vpi: Social Data Commons: Education & Training: Virginia Preschool Initiative Access}
\description{
Allows user to easily get VPI provider accees data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
